+++ lib/rabl-rails/library.rb
@@ -14,8 +14,9 @@ module RablRails
 
       compiled_template = compile_template_from_source(source, path)
 
-      format = context.params[:format] || 'json'
-      Renderers.const_get(format.to_s.upcase!).new(context, locals).render(compiled_template)
+      format = context.params[:format] ? context.params[:format].to_s : 'json'
+      format.upcase!
+      Renderers.const_get(format).new(context, locals).render(compiled_template)
     end
 
     def compile_template_from_source(source, path = nil)